+++ Sources/Servers/Giny.World/Managers/Fights/Effects/Debuffs/StatsDebuffPercent.cs
@@ -23,7 +23,7 @@ namespace Giny.World.Managers.Fights.Effects.Debuffs
 
         protected override void Apply(IEnumerable<Fighter> targets)
         {
-            short delta = Effect.GetDelta();
+            int delta = Effect.GetDelta();
 
             foreach (var target in targets)
             {
